Refactored method on createPlayersServices

+++ src/players/services/deletePlayer.service.ts
@@ -1,4 +1,4 @@
-import { Injectable } from '@nestjs/common';
+import { Injectable, NotFoundException } from '@nestjs/common';
 import { InjectModel } from '@nestjs/mongoose';
 import { Model } from 'mongoose';
 import { Player } from '../interfaces/palyer.interface';
@@ -10,6 +10,12 @@ export class DeletePlayerService {
   ) {}
 
   async execute(Id: string): Promise<void> {
+    const player = await this.playerModel.findById(Id);
+
+    if (!player) {
+      throw new NotFoundException('Player not found');
+    }
+
     await this.playerModel.findByIdAndDelete(Id).exec();
   }
 }
